Nest breviloquence CborParser as Cbor.Parser
Move the private CborParser object+class into object Cbor as Cbor.Parser,
deleting the standalone file. Purely an internal reorganisation — the
parser was already private[breviloquence] and not in the soundness
surface. breviloquence compiles; 65 tests pass.
