+++ lib/breviloquence/src/core/breviloquence_parser.scala
@@ -38,7 +38,7 @@ import prepositional.*
 import turbulence.*
 
 extension (cbor: Cbor.Ast.type)
-  def parse(source: IArray[Byte]): Cbor.Ast raises CborError = CborParser.parse(source)
+  def parse(source: IArray[Byte]): Cbor.Ast raises CborError = Cbor.Parser.parse(source)
 
 given parserAggregable: Tactic[CborError] => Cbor.Ast is Aggregable by Data =
   source => Cbor.Ast.parse(source.read[Data])
